Allow selecting specific groups to add entries for
If you choose to add Ansible managed hosts, it will include all
hosts which may be unnecessary in large deployments.  The newly
added variable `hosts_add_ansible_managed_hosts_groups` allows
the user to select which specific groups they would like to add.

The default value is set to ['all'] in order to maintain the
same functionality as before.